02_quant_research/02_cointegration_stat_arb_kalman_spreads/requirements.txt
numpy>=1.26.4,<2.0.0
pandas>=2.0.3
scipy>=1.11.1
statsmodels>=0.14.0
matplotlib>=3.7.2
seaborn>=0.12.2
yfinance>=0.2.40
pyyaml>=6.0.1
scikit-learn>=1.3.0
pykalman>=0.9.7
jupyter>=1.0.0
notebook>=7.0.2
tqdm>=4.66.1
